Fixed #23 -- Handle the existence of TestSuites in build_suite() output.
This generally won't happen, but on Django 1.5, when unittest2 is manually
installed, there is an edge case bug in Django's test suite reordering that
causes empty test suites to be added to the output of build_suite().

More generally, it's possible for a TestSuites to be recursive, so this will
account for that.

Thanks to @lensvol for the report, debug help and draft patch.